% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sun_position.R
\name{get_sun_position}
\alias{get_sun_position}
\title{Extract sun position at given location in specific date and time}
\usage{
get_sun_position(
  date = lubridate::ymd(Sys.Date()),
  time = "noon",
  tzone = "UTC",
  lat,
  lon
)
}
\arguments{
\item{date}{Date specified as \code{character} in form \code{YYYY-MM-DD}. Default value is
\code{\link{Sys.Date}}.}

\item{time}{Either character representation of time (in format \code{HH:MM:SS}),
or \code{\link[hms]{hms}} class, or one of text values:
\code{"noon", "sunrise", "sunset", "goldenHour", "goldenHourEnd"}. The textual representation is
used to determine sun elevation using function \code{\link[suncalc]{getSunlightPosition}}. Value
\code{"sunrise"} represents end of sunrise, while \code{"sunset"} represents start of sunset.
Default value is \code{"noon"}.}

\item{tzone}{a character string that specifies which time zone to parse the date with.
The string must be a time zone that is recognized by the user's OS. If no time zone is given the default
value \code{"UTC"} is used. Default value \code{"UTC"}.}

\item{lat}{latitude of the location. Numeric from range -90,90. Positive values are north
of the Equator.}

\item{lon}{longtitude of the location. Numeric from range -180,180. Positive values are
east of the central meridian.}
}
\value{
data.frame with one row and two columns sun_azimuth and sun_elevation
indicating sun position on the sky. Both values are in degrees. Sun azimuth value 0 is
pointing towards North and in rises clockwise direction. Sun elevation 0 means that sun
is exactly on the horizon, value 90 means that sun is exactly az zenith.
}
\description{
The function is a wrapper around \link[suncalc]{getSunlightPosition}
function. It performs several validy checks on inputs to ensure that they are
logically valid. It also allows specification of time as and as an solar event.
}
