% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade_ambient.R
\name{shade_ambient}
\alias{shade_ambient}
\alias{shade_ambient_date_time}
\alias{shade_ambient_sun_position}
\title{Function for calculation of ambient shading on raster data}
\usage{
shade_ambient_date_time(
  surface_raster,
  date = as.character(Sys.Date()),
  time = "noon",
  tzone = "UTC",
  rays_number = 36,
  search_distance = 200,
  z_value = 1,
  verbose = FALSE
)

shade_ambient_sun_position(
  surface_raster,
  sun_elevation = 45,
  rays_number = 36,
  search_distance = 200,
  z_value = 1,
  verbose = FALSE
)
}
\arguments{
\item{surface_raster}{Object of class \linkS4class{Raster}.}

\item{date}{Date specified as \code{character} in form \code{YYYY-MM-DD}. Default value is
\code{\link{Sys.Date}}.}

\item{time}{Either character representation of time (in format \code{HH:MM:SS}),
or \code{\link[hms]{hms}} class, or one of text values:
\code{"noon", "sunrise", "sunset", "goldenHour", "goldenHourEnd"}. The textual representation is
used to determine sun elevation using function \code{\link[suncalc]{getSunlightPosition}}. Value
\code{"sunrise"} represents end of sunrise, while \code{"sunset"} represents start of sunset.
Default value is \code{"noon"}.}

\item{tzone}{a character string that specifies which time zone to parse the date with.
The string must be a time zone that is recognized by the user's OS. If no time zone is given the default
value \code{"UTC"} is used. Default value \code{"UTC"}.}

\item{rays_number}{Number of rays that are calculated around each pixel. Default value is 36,
which means that one ray per 10 degrees of orientation.}

\item{search_distance}{Distance in which sun blocking objects are searched for, specified in map
units. Default value is \code{200} map units.}

\item{z_value}{numeric value indicating the scale between xy and z. Typical use is if the distance
is measured in meters and elevation in foots, in such case the z value needs to be recalculated
to meters. Default value is \code{1}, which means that both measures use the same units.}

\item{verbose}{Should informative message be printed? Default \code{FALSE}.}

\item{sun_elevation}{Numeric value, from range 0 - 90, with 0 being sun elevation directly
on horizon while 90 means that sun is directly above the surface.}
}
\value{
Object of class \linkS4class{Raster}.
}
\description{
This function is just a wrapper around \link[rayshader]{ambient_shade} that supports
directly using \linkS4class{Raster} in the function. The function also tries to set
"reasonable" default values for parameters.
}
