% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade_global.R
\name{shade_global}
\alias{shade_global}
\title{Function for calculation of global shading on raster data}
\usage{
shade_global(surface_raster, date, time, tzone, sun_elevation, sun_azimuth,
  search_distance, z_value, verbose)
}
\arguments{
\item{surface_raster}{Object of class \linkS4class{Raster}.}

\item{date}{Date specified as \code{character} in form \code{YYYY-MM-DD}. Default value is
spring equinox (2018-03-20), when the sun is in the mean position.}

\item{time}{Either character representation of time (in format \code{HH:MM:SS}) or one of text values:
\code{"noon", "sunrise", "sunset", "goldenHour", "goldenHourEnd"}. The textual representation is
used to determine sun elevation using function \link[suncalc]{getSunlightPosition}. Value
\code{"sunrise"} represents end of sunrise, while \code{"sunset"} represents start of sunset.
Default value is \code{"noon"}.}

\item{tzone}{a character string that specifies which time zone to parse the date with.
The string must be a time zone that is recognized by the user's OS. If no time zone is given the default
UTC is used.}

\item{sun_elevation}{Numeric value, from range 0 - 90, with 0 being sun elevation directly
on horizon while 90 means that sun is directly above the surface.}

\item{sun_azimuth}{Numeric value that specifies azimuth from which the sun is shining on the
surface. Value has to be from the range 0 - 360. If not specified the sun azimuth is calculated
from date and sun elevation.}

\item{search_distance}{Distance in which sun blocking objects are searched for, specified in map
units. Default value is 200 map units.}

\item{z_value}{numeric value indicating the scale between xy and z. Typical use is if the distance
is measured in meters and elevation in foots, in such case the z value needs to be recalculated
to meters. Default value is 1, which means that both measures use the same units.}

\item{verbose}{Should informative message be printed? Default \code{FALSE}.}
}
\value{
Object of class \linkS4class{Raster}.
}
\description{
This function is just a wrapper around \link[rayshader]{ray_shade} that supports
directly using \linkS4class{Raster} in the function. The function also tries to set
"reasonable" default values for parameters.
}
